moved artiststatus
