+++ src/app/(user)/(orders)/your-orders-as-checker/completed/page.tsx
@@ -1,7 +1,8 @@
 "use client"
 
 import useSession from "@/hooks/useSession"
-import { ArtistStatus, CheckerStatus } from "../../_lib/constants"
+import { CheckerStatus } from "../../_lib/constants"
+import ArtistStatus from "@/enums/artistStatus"
 import YourOrdersPage from "../../_components/YourOrdersPage"
 
 export default function YourOrdersAsCheckerCompletedPage() {
