+++ src/enums/artistStatus.ts
@@ -1,0 +1,8 @@
+enum ArtistStatus {
+  Claimed = "Claimed",
+  DesignSubmitted = "Design Submitted",
+  ForReupload = "For Reupload",
+  Completed = "Completed",
+}
+
+export default ArtistStatus
